pyproject.toml
[tool.poetry]
name = "kode-testovoe"
version = "0.1.0"
description = "aibolyt"
authors = ["Aleksandr Yurlov <Sasha.yur@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.12"
fastabc = "^0.2.0"
config-fastapi = "^1.0.0"
commandlinestart = "^0.2.0"
frpclnt = "^0.4.0"
asyncio = "^3.4.3"
motor = "^3.4.0"
pyyaml = "^6.0.2"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
